Unimportant check in (removed whitespaces etc)

+++ src/GeomCompactionManager.cpp
@@ -36,13 +36,13 @@ GeomCompactionManager::GeomCompactionManager(MemStore *memstore, DiskStore *disk
         fscanf(fp, "R: %d\n", &m_R);
         fscanf(fp, "P: %d\n", &m_P);
         fscanf(fp, "partitions: %d\n", &num_partitions);
-        for (uint i = 0; i < num_partitions; i++) {
+        for (int i = 0; i < num_partitions; i++) {
             fscanf(fp, "%d\n", &part_size);
             m_partition_size.push_back(part_size);
             dbg_lastsize += part_size;
         }
 
-        assert(m_partition_size.size() == num_partitions);
+        assert((int)m_partition_size.size() == num_partitions);
         sanity_check();
 
         fclose(fp);
@@ -68,7 +68,7 @@ GeomCompactionManager::~GeomCompactionManager()
         fprintf(fp, "cmmanager: %s\n", "geometric");
         fprintf(fp, "R: %d\n", m_R);
         fprintf(fp, "P: %d\n", m_P);
-        fprintf(fp, "partitions: %d\n", m_partition_size.size());
+        fprintf(fp, "partitions: %d\n", (int)m_partition_size.size());
         for (uint i = 0; i < m_partition_size.size(); i++) {
             fprintf(fp, "%d\n", m_partition_size[i]);
         }
